Improve sonar coverage
