+++ pkg/hub/http_user_server.go
@@ -9,7 +9,6 @@ import (
 	"net/http/httputil"
 	"net/url"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/spf13/cobra"
@@ -165,18 +164,6 @@ func (k *HTTPUserServer) handler(wr http.ResponseWriter, req *http.Request) {
 	klog.V(4).Infof("request scheme:%s; rawQuery:%s; path:%s", req.URL.Scheme, req.URL.RawQuery, req.URL.Path)
 
 	proxy.ServeHTTP(wr, req)
-}
-
-func parseRequestURL(requestURL string) (clusterID string, kubeAPIPath string, err error) {
-	paths := strings.Split(requestURL, "/")
-	if len(paths) <= 2 {
-		err = fmt.Errorf("requestURL format not correct, path more than 2: %s", requestURL)
-		return
-	}
-	clusterID = paths[1]                             // <clusterID>
-	kubeAPIPath = strings.Join(paths[2:], "/")       // api/pods?timeout=32s
-	kubeAPIPath = strings.Split(kubeAPIPath, "?")[0] // api/pods note: we only need path here, the proxy pkg would add params back
-	return
 }
 
 func (k *HTTPUserServer) Run(ctx context.Context) error {
