+++ pkg/hub/parse.go
@@ -1,0 +1,18 @@
+package hub
+
+import (
+	"fmt"
+	"strings"
+)
+
+func parseRequestURL(requestURL string) (clusterID string, kubeAPIPath string, err error) {
+	paths := strings.Split(requestURL, "/")
+	if len(paths) <= 2 {
+		err = fmt.Errorf("requestURL format not correct, path more than 2: %s", requestURL)
+		return
+	}
+	clusterID = paths[1]                             // <clusterID>
+	kubeAPIPath = strings.Join(paths[2:], "/")       // api/pods?timeout=32s
+	kubeAPIPath = strings.Split(kubeAPIPath, "?")[0] // api/pods note: we only need path here, the proxy pkg would add params back
+	return
+}
