+++ sonar-project.properties
@@ -4,7 +4,7 @@ sonar.sources=.
 sonar.exclusions=**/*_test.go,**/vendor/**,/test/**
 sonar.tests=.
 sonar.test.inclusions=**/*_test.go
-sonar.test.exclusions=**/vendor/**,/test/**
+sonar.test.exclusions=**/vendor/**,/test/**,**/main.go,**/version.go,**/http_user_server.go,**/cmd/**
 sonar.go.tests.reportPaths=report.json
 sonar.go.coverage.reportPaths=coverage.out
 sonar.externalIssuesReportPaths=gosec.json
